Move a bunch of existing code to YouRobotManager
